labtoolIX/requirements.txt
Django==1.5
South==0.7.6
django-auth-ldap==1.1.2
django-tastypie==0.10
django-durationfield==0.4.0
mimeparse==0.1.3
pyasn1==0.1.6
pycrypto==2.6
pysnmp==4.2.4
python-dateutil==1.5
python-ldap==2.4.10
splinter==0.5.2
ssh==1.8.0
wsgiref==0.1.2
